chore: move banners under shared/components/marketing
Pure folder rename: shared/components/banners -> shared/components/marketing/banners.
Updates 9 import sites and the relative imports within moved files. No behavior change.

First step of consolidating marketing surfaces under a single shared/components/marketing/ parent.
