+++ packages/shared/src/components/MainFeedLayout.tsx
@@ -11,8 +11,8 @@ import classNames from 'classnames';
 import { useRouter } from 'next/router';
 import type { FeedProps } from './Feed';
 import Feed from './Feed';
-import ReadingReminderHero from './banners/ReadingReminderHero';
-import ShortcutsExtensionPromo from './banners/ShortcutsExtensionPromo';
+import ReadingReminderHero from './marketing/banners/ReadingReminderHero';
+import ShortcutsExtensionPromo from './marketing/banners/ShortcutsExtensionPromo';
 import { WebappShortcutsRow } from '../features/shortcuts/components/WebappShortcutsRow';
 import { AskSearchBanner } from './notifications/AskSearchBanner';
 import AuthContext from '../contexts/AuthContext';
