+++ packages/shared/src/components/marketing/banners/HeroBottomBanner.tsx
@@ -1,9 +1,9 @@
 import classNames from 'classnames';
 import type { ReactElement } from 'react';
 import React from 'react';
-import { Button, ButtonVariant } from '../buttons/Button';
-import { MiniCloseIcon } from '../icons';
-import feedStyles from '../Feed.module.css';
+import { Button, ButtonVariant } from '../../buttons/Button';
+import { MiniCloseIcon } from '../../icons';
+import feedStyles from '../../Feed.module.css';
 import ReadingReminderCatLaptop from './ReadingReminderCatLaptop';
 
 type TopHeroProps = {
