+++ packages/shared/src/components/marketing/banners/common.ts
@@ -1,4 +1,4 @@
-import classed from '../../lib/classed';
+import classed from '../../../lib/classed';
 
 export const bottomBannerBaseClasses =
   'fixed bottom-0 left-0 z-modal flex w-full flex-row justify-center border-none';
